Rewriting the Notify-Low-Battery Script
The script I wrote before does not work, so I rewrote it.

Also it is in my opinion a little bit better, to use watch, then notify-wait. Because
instead of checking if the file was changed and check the /sys/class/BAT1/capacity
file for modifications every 5 minutes isn't as efficient as simply using watch to
check the battery status every 5 minutes.